Localized JS strings
Added localization for strings which need to be translated immediately
in JS before reaching the DOM. This is done in edge cases where it's not
desired to inject arbitrary HTML code to accommodate L20n philosophy of
using data-l10n-id attributes on HTML elements. Rather than letting L20n
translate these strings in the DOM, they are translated in JS first.